number-euclid: large numbers are here very usual

+++ code/number-euclid.cpp
@@ -3,8 +3,8 @@
 // http://atrey.karlin.mff.cuni.cz/~folwar/insalg/
 
 /*pdf*/
-int euclid(int a, int b) {
-	int t;
+ll euclid(ll a, ll b) {
+	ll t;
 	while (b != 0)  {
 		t = b;
 		b = a % b;
@@ -14,13 +14,13 @@ int euclid(int a, int b) {
 }
 
 //GCD = a * x + b * y
-int euclid_extended(int a, int b, int &x, int &y) {
+ll euclid_extended(ll a, ll b, ll &x, ll &y) {
 	if (b == 0) {
 		x = 1; y = 0;
 		return a;
 	}
-	int xx, yy;
-	int gcd = euclid_extended(b, a%b, xx, yy);
+	ll xx, yy;
+	ll gcd = euclid_extended(b, a%b, xx, yy);
 	x = yy;
 	y = xx - (a/b) * yy;
 	return gcd;
@@ -29,15 +29,15 @@ int euclid_extended(int a, int b, int &x, int &y) {
 
 // DEMO
 void euclid_demo() {
-	printf("GCD(22031993, 2203) = %d\n", euclid(22031993,2203));
-	printf("GCD(332024, 224224) = %d\n", euclid(332024, 224224));
-	int a = 13, b = 17, x, y;
-	int gcd = euclid_extended(a, b, x, y);
-	printf("%d = %d * %d + %d * %d\n", gcd, a, x, b, y);
+	printf("GCD(22031993, 2203) = %lld\n", euclid(22031993,2203));
+	printf("GCD(332024, 224224) = %lld\n", euclid(332024, 224224));
+	ll a = 13, b = 17, x, y;
+	ll gcd = euclid_extended(a, b, x, y);
+	printf("%lld = %lld * %lld + %lld * %lld\n", gcd, a, x, b, y);
 
 	a = 8, b = 24;
 	gcd = euclid_extended(a, b, x, y);
-	printf("%d = %d * %d + %d * %d\n", gcd, a, x, b, y);
+	printf("%lld = %lld * %lld + %lld * %lld\n", gcd, a, x, b, y);
 
 }
